Added updateOrCreate function
One of the great functions that Eloquent has is to be able to create
a model if it doesn't exist, or update it if it does.

Calling $model->updateOrCreate($attributes, $values) will search for a
model matching the values in $attributes and if found will update it
with the values in $values. If it wasn't found, then it will create a
model for it. In both situations the instance of that model will be
returned.